feat(messages,pretty): add event and service check parsing and styling
Add DogStatsD event and service check parsing to messages package.
Move metric, event, and service check pretty-formatting to separate
files.
Implement styling helpers for events and service checks in pretty
output.
